Revert "Merge pull request #5 from TRON-US/BTFS-1401"
This reverts commit 0f7370ce12b111ff514ef0e05efe6cbe98e28288, reversing
changes made to 96a7168d291ade7bf9f937093dbd5a6fc3c97f3d.
