--- utils/peerId.go
@@ -1,5 +1,0 @@
-package utils
-
-func generatePeerId(publicKey string) (string) {
-	return "16Uiu2HAkwQZvY1mQjWabNr6eDKR7SW5i1KsnVyQpZqWmtRJ6u6SB"
-}
